rename layers in projects
